requirements.txt
python-dotenv
langchain==0.3.27
langchain-openai
langchain-anthropic
langchain-together
numpy>=1.24,<2.0
sentence-transformers
grpcio==1.47.0
langchain-google-genai